R Data List of Tables
Is a list of tables, each table presents the ordered genes along the DNA
strand. For each gene is presented the locus name, the coding
sequence(label cds), the codon adaptation index(label CAI), the
direction 5'-3'\3'-5'(label senso), the symbol of the associated protein
(label genesym), the gene length (label genel), the protein abundance in
standard condition (label PA). Eeach object of the list have a name that
corresponds to the KEGG pathway identifier for the organism "eco".